fix: movie_id로 통일

+++ app/models/comment.py
@@ -1,6 +1,6 @@
 # app/models/comment.py
 
-from sqlalchemy import Column, BigInteger, Text, Boolean, DECIMAL, DateTime, ForeignKey
+from sqlalchemy import Column, BigInteger, Integer, Text, Boolean, DECIMAL, DateTime, ForeignKey
 from sqlalchemy.sql import func
 from app.database import Base
 
@@ -8,7 +8,7 @@ class CommentModel(Base):
     __tablename__ = "comments"
     
     comment_id = Column(BigInteger, primary_key=True, autoincrement=True)
-    movie_id = Column(BigInteger, ForeignKey('movies.movie_id'), nullable=False)
+    movie_id = Column(Integer, ForeignKey('movies.movie_id'), nullable=False)
     user_id = Column(BigInteger, ForeignKey('users.user_id'), nullable=False)
     content = Column(Text, nullable=False)
     is_spoiler = Column(Boolean, default=False)
